Move most of the transfer usecases and workers to a specific package to keep them organised
