+++ owncloudApp/src/main/java/com/owncloud/android/ui/activity/ConflictsResolveActivity.java
@@ -28,9 +28,9 @@ import com.owncloud.android.domain.files.model.OCFile;
 import com.owncloud.android.ui.dialog.ConflictsResolveDialog;
 import com.owncloud.android.ui.dialog.ConflictsResolveDialog.Decision;
 import com.owncloud.android.ui.dialog.ConflictsResolveDialog.OnConflictDecisionMadeListener;
-import com.owncloud.android.usecases.UploadFileInConflictUseCase;
-import com.owncloud.android.usecases.UploadFilesFromSystemUseCase;
-import com.owncloud.android.usecases.transfers.DownloadFileUseCase;
+import com.owncloud.android.usecases.transfers.uploads.UploadFileInConflictUseCase;
+import com.owncloud.android.usecases.transfers.uploads.UploadFilesFromSystemUseCase;
+import com.owncloud.android.usecases.transfers.downloads.DownloadFileUseCase;
 import kotlin.Lazy;
 import org.jetbrains.annotations.NotNull;
 import timber.log.Timber;
