+++ owncloudApp/src/main/java/com/owncloud/android/ui/activity/ManageAccountsActivity.java
@@ -26,7 +26,6 @@ import android.accounts.AccountManagerCallback;
 import android.accounts.AccountManagerFuture;
 import android.accounts.OperationCanceledException;
 import android.content.ContentResolver;
-import android.content.Context;
 import android.content.Intent;
 import android.content.SyncRequest;
 import android.graphics.drawable.Drawable;
@@ -52,8 +51,8 @@ import com.owncloud.android.ui.adapter.AccountListItem;
 import com.owncloud.android.ui.dialog.RemoveAccountDialogFragment;
 import com.owncloud.android.ui.dialog.RemoveAccountDialogViewModel;
 import com.owncloud.android.ui.helpers.FileOperationsHelper;
-import com.owncloud.android.usecases.CancelUploadFromAccountUseCase;
-import com.owncloud.android.usecases.transfers.CancelDownloadsForAccountUseCase;
+import com.owncloud.android.usecases.transfers.uploads.CancelUploadFromAccountUseCase;
+import com.owncloud.android.usecases.transfers.downloads.CancelDownloadsForAccountUseCase;
 import com.owncloud.android.utils.PreferenceUtils;
 import kotlin.Lazy;
 import org.jetbrains.annotations.NotNull;
