+++ owncloudApp/src/main/java/com/owncloud/android/ui/adapter/ExpandableUploadListAdapter.java
@@ -55,9 +55,9 @@ import com.owncloud.android.lib.common.network.OnDatatransferProgressListener;
 import com.owncloud.android.ui.activity.FileActivity;
 import com.owncloud.android.ui.fragment.OptionsInUploadListClickListener;
 import com.owncloud.android.ui.fragment.UploadListFragment;
-import com.owncloud.android.usecases.CancelUploadWithIdUseCase;
-import com.owncloud.android.usecases.RetryUploadFromContentUriUseCase;
-import com.owncloud.android.usecases.RetryUploadFromSystemUseCase;
+import com.owncloud.android.usecases.transfers.uploads.CancelUploadWithIdUseCase;
+import com.owncloud.android.usecases.transfers.uploads.RetryUploadFromContentUriUseCase;
+import com.owncloud.android.usecases.transfers.uploads.RetryUploadFromSystemUseCase;
 import com.owncloud.android.utils.DisplayUtils;
 import com.owncloud.android.utils.MimetypeIconUtil;
 import com.owncloud.android.utils.PreferenceUtils;
