+++ owncloudApp/src/main/java/com/owncloud/android/ui/helpers/FileOperationsHelper.java
@@ -41,8 +41,8 @@ import com.owncloud.android.presentation.ui.sharing.ShareActivity;
 import com.owncloud.android.services.OperationsService;
 import com.owncloud.android.ui.activity.FileActivity;
 import com.owncloud.android.ui.dialog.ShareLinkToDialog;
-import com.owncloud.android.usecases.CancelUploadForFileUseCase;
-import com.owncloud.android.usecases.transfers.CancelDownloadForFileUseCase;
+import com.owncloud.android.usecases.transfers.uploads.CancelUploadForFileUseCase;
+import com.owncloud.android.usecases.transfers.downloads.CancelDownloadForFileUseCase;
 import com.owncloud.android.utils.UriUtilsKt;
 import kotlin.Lazy;
 import org.jetbrains.annotations.NotNull;
