+++ owncloudApp/src/main/java/com/owncloud/android/ui/preview/FileDownloadFragment.kt
@@ -38,7 +38,7 @@ import com.owncloud.android.R
 import com.owncloud.android.domain.files.model.OCFile
 import com.owncloud.android.extensions.observeWorkerTillItFinishes
 import com.owncloud.android.ui.fragment.FileFragment
-import com.owncloud.android.usecases.transfers.GetLiveDataForDownloadingFileUseCase
+import com.owncloud.android.usecases.transfers.downloads.GetLiveDataForDownloadingFileUseCase
 import com.owncloud.android.utils.PreferenceUtils
 import org.koin.android.ext.android.inject
