+++ owncloudApp/src/main/java/com/owncloud/android/usecases/transfers/downloads/GetLiveDataForDownloadingFileUseCase.kt
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.owncloud.android.usecases.transfers
+package com.owncloud.android.usecases.transfers.downloads
 
 import android.accounts.Account
 import androidx.lifecycle.LiveData
